bigtable: move manually edited types out of generated file
